refactor: split out TEColorParameter, TECommonAngleParameter
